Refactor evidence API tests and remove stale code
