+++ services/artana_evidence_api/document_extraction_relation_taxonomy.py
@@ -1,0 +1,272 @@
+"""LLM relation taxonomy for document extraction."""
+
+from __future__ import annotations
+
+# Canonical relation types matching GRAPH_SERVICE_BUILTIN_RELATION_TYPES.
+# This set is validated against the graph domain config by
+# test_constraint_profile_model.py::test_llm_valid_types_matches_builtin_types
+# to ensure it stays in sync.  We cannot import from artana_evidence_db
+# directly due to the service boundary.
+LLM_VALID_RELATION_TYPES = frozenset(
+    {
+        # Core causal
+        "ASSOCIATED_WITH",
+        "CAUSES",
+        "TREATS",
+        "TARGETS",
+        "BIOMARKER_FOR",
+        "PHYSICALLY_INTERACTS_WITH",
+        "ACTIVATES",
+        "REGULATES",
+        "INHIBITS",
+        "SENSITIZES_TO",
+        "PHENOCOPY_OF",
+        # Extended scientific
+        "UPSTREAM_OF",
+        "DOWNSTREAM_OF",
+        "PART_OF",
+        "COMPONENT_OF",
+        "EXPRESSED_IN",
+        "PARTICIPATES_IN",
+        "LOCATED_IN",
+        # Mechanistic / functional consequence
+        "MODULATES",
+        "LOSS_OF_FUNCTION",
+        "GAIN_OF_FUNCTION",
+        "PREDISPOSES_TO",
+        "CO_OCCURS_WITH",
+        "COLOCALIZES_WITH",
+        "COMPENSATED_BY",
+        "SUBSTRATE_OF",
+        "TRANSPORTS",
+        # Evidence / governance
+        "SUPPORTS",
+        "REFINES",
+        "GENERALIZES",
+        "INSTANCE_OF",
+        "MENTIONS",
+        "CITES",
+        "HAS_AUTHOR",
+        "HAS_KEYWORD",
+    },
+)
+
+# Synonyms that map to canonical types (from graph_domain_config.py)
+LLM_RELATION_SYNONYMS: dict[str, str] = {
+    # ASSOCIATED_WITH
+    "ASSOCIATES_WITH": "ASSOCIATED_WITH",
+    "LINKED_TO": "ASSOCIATED_WITH",
+    "CORRELATED_WITH": "ASSOCIATED_WITH",
+    "CORRELATES_WITH": "ASSOCIATED_WITH",
+    "RELATED_TO": "ASSOCIATED_WITH",
+    "CONNECTED_TO": "ASSOCIATED_WITH",
+    "IMPLICATED_IN": "ASSOCIATED_WITH",
+    "TIED_TO": "ASSOCIATED_WITH",
+    "OBSERVED_IN": "ASSOCIATED_WITH",
+    "REPORTED_IN": "ASSOCIATED_WITH",
+    # CAUSES
+    "LEADS_TO": "CAUSES",
+    "RESULTS_IN": "CAUSES",
+    "INDUCES": "CAUSES",
+    "PRODUCES": "CAUSES",
+    "CONTRIBUTES_TO": "CAUSES",
+    "CONFERS": "CAUSES",
+    "GIVES_RISE_TO": "CAUSES",
+    "TRIGGERS": "CAUSES",
+    "ELICITS": "CAUSES",
+    "DRIVES": "CAUSES",
+    "ENGENDERS": "CAUSES",
+    # TREATS
+    "THERAPEUTIC_FOR": "TREATS",
+    "AMELIORATES": "TREATS",
+    "ALLEVIATES": "TREATS",
+    "USED_TO_TREAT": "TREATS",
+    "REVERSES": "TREATS",
+    "CURES": "TREATS",
+    "MANAGES": "TREATS",
+    "MITIGATES": "TREATS",
+    "RESOLVES": "TREATS",
+    # TARGETS
+    "ACTS_ON": "TARGETS",
+    "BINDS_TO": "TARGETS",
+    "DIRECTED_AT": "TARGETS",
+    "ENGAGES": "TARGETS",
+    "ATTACKS": "TARGETS",
+    "ACTS_AGAINST": "TARGETS",
+    # BIOMARKER_FOR
+    "PREDICTS": "BIOMARKER_FOR",
+    "DIAGNOSTIC_FOR": "BIOMARKER_FOR",
+    "PROGNOSTIC_FOR": "BIOMARKER_FOR",
+    "INDICATIVE_OF": "BIOMARKER_FOR",
+    "MARKER_FOR": "BIOMARKER_FOR",
+    "PREDICTOR_OF": "BIOMARKER_FOR",
+    "INDICATOR_OF": "BIOMARKER_FOR",
+    "READ_OUT_FOR": "BIOMARKER_FOR",
+    # PHYSICALLY_INTERACTS_WITH
+    "INTERACTS_WITH": "PHYSICALLY_INTERACTS_WITH",
+    "BINDS": "PHYSICALLY_INTERACTS_WITH",
+    "COMPLEXES_WITH": "PHYSICALLY_INTERACTS_WITH",
+    "DIMERIZES_WITH": "PHYSICALLY_INTERACTS_WITH",
+    "FORMS_COMPLEX_WITH": "PHYSICALLY_INTERACTS_WITH",
+    "INTERACTS_PHYSICALLY_WITH": "PHYSICALLY_INTERACTS_WITH",
+    "DOCKS_WITH": "PHYSICALLY_INTERACTS_WITH",
+    # ACTIVATES
+    "STIMULATES": "ACTIVATES",
+    "PROMOTES": "ACTIVATES",
+    "ENHANCES": "ACTIVATES",
+    "UPREGULATES": "ACTIVATES",
+    "INDUCES_EXPRESSION_OF": "ACTIVATES",
+    "TURNS_ON": "ACTIVATES",
+    "INITIATES": "ACTIVATES",
+    "SWITCHES_ON": "ACTIVATES",
+    "AUGMENTS": "ACTIVATES",
+    # INHIBITS
+    "SUPPRESSES": "INHIBITS",
+    "REPRESSES": "INHIBITS",
+    "BLOCKS": "INHIBITS",
+    "DOWNREGULATES": "INHIBITS",
+    "ATTENUATES": "INHIBITS",
+    "SILENCES": "INHIBITS",
+    "DAMPENS": "INHIBITS",
+    "ANTAGONIZES": "INHIBITS",
+    "ABROGATES": "INHIBITS",
+    "ABOLISHES": "INHIBITS",
+    # REGULATES
+    "MEDIATES": "REGULATES",
+    "CONTROLS": "REGULATES",
+    "GOVERNS": "REGULATES",
+    "MODULATES_ACTIVITY_OF": "MODULATES",
+    "INFLUENCES": "REGULATES",
+    "ORCHESTRATES": "REGULATES",
+    "COORDINATES": "REGULATES",
+    # EXPRESSED_IN
+    "EXPRESSED_WITHIN": "EXPRESSED_IN",
+    "EXPRESSED_BY": "EXPRESSED_IN",
+    "SHOWS_EXPRESSION_IN": "EXPRESSED_IN",
+    "DETECTED_IN": "EXPRESSED_IN",
+    "PRESENT_IN": "EXPRESSED_IN",
+    "TRANSCRIBED_IN": "EXPRESSED_IN",
+    "ENRICHED_IN": "EXPRESSED_IN",
+    "ABUNDANT_IN": "EXPRESSED_IN",
+    # PARTICIPATES_IN
+    "INVOLVED_IN": "PARTICIPATES_IN",
+    "FUNCTIONS_IN": "PARTICIPATES_IN",
+    "TAKES_PART_IN": "PARTICIPATES_IN",
+    "PLAYS_ROLE_IN": "PARTICIPATES_IN",
+    "ACTIVE_IN": "PARTICIPATES_IN",
+    "OPERATES_IN": "PARTICIPATES_IN",
+    # LOCATED_IN
+    "LOCALIZES_TO": "LOCATED_IN",
+    "FOUND_IN": "LOCATED_IN",
+    "RESIDES_IN": "LOCATED_IN",
+    "LOCALIZED_TO": "LOCATED_IN",
+    "PRESENT_AT": "LOCATED_IN",
+    "SITUATED_IN": "LOCATED_IN",
+    "CONFINED_TO": "LOCATED_IN",
+    # PREDISPOSES_TO
+    "RISK_FACTOR_FOR": "PREDISPOSES_TO",
+    "INCREASES_RISK_OF": "PREDISPOSES_TO",
+    "SUSCEPTIBILITY_TO": "PREDISPOSES_TO",
+    "CONFERS_SUSCEPTIBILITY": "PREDISPOSES_TO",
+    "ELEVATES_RISK_OF": "PREDISPOSES_TO",
+    "PREDISPOSES": "PREDISPOSES_TO",
+    # CO_OCCURS_WITH
+    "COINCIDES_WITH": "CO_OCCURS_WITH",
+    "COMORBID_WITH": "CO_OCCURS_WITH",
+    "CO_PRESENTS_WITH": "CO_OCCURS_WITH",
+    "COEXISTS_WITH": "CO_OCCURS_WITH",
+    "FOUND_TOGETHER_WITH": "CO_OCCURS_WITH",
+    # UPSTREAM_OF / DOWNSTREAM_OF
+    "PRECEDES": "UPSTREAM_OF",
+    "SIGNALS_TO": "UPSTREAM_OF",
+    "FEEDS_INTO": "UPSTREAM_OF",
+    "ACTS_BEFORE": "UPSTREAM_OF",
+    "FOLLOWS": "DOWNSTREAM_OF",
+    "TRIGGERED_BY": "DOWNSTREAM_OF",
+    "RECEIVES_SIGNAL_FROM": "DOWNSTREAM_OF",
+    "ACTS_AFTER": "DOWNSTREAM_OF",
+    # PART_OF / COMPONENT_OF
+    "SUBUNIT_OF": "PART_OF",
+    "CONTAINED_IN": "PART_OF",
+    "MEMBER_OF": "PART_OF",
+    "BELONGS_TO": "PART_OF",
+    "INCLUDED_IN": "PART_OF",
+    "INTEGRAL_PART_OF": "PART_OF",
+    "CONSTITUENT_OF": "COMPONENT_OF",
+    "ELEMENT_OF": "COMPONENT_OF",
+    "BUILDING_BLOCK_OF": "COMPONENT_OF",
+    # SENSITIZES_TO
+    "INCREASES_SENSITIVITY_TO": "SENSITIZES_TO",
+    "CONFERS_SENSITIVITY_TO": "SENSITIZES_TO",
+    "RENDERS_SENSITIVE_TO": "SENSITIZES_TO",
+    # PHENOCOPY_OF
+    "PHENOTYPICALLY_MIMICS": "PHENOCOPY_OF",
+    "CLINICALLY_INDISTINGUISHABLE_FROM": "PHENOCOPY_OF",
+    "MIMICS": "PHENOCOPY_OF",
+    # COLOCALIZES_WITH
+    "CO_LOCALIZES_WITH": "COLOCALIZES_WITH",
+    "FOUND_WITH": "COLOCALIZES_WITH",
+    "CO_EXPRESSED_WITH": "COLOCALIZES_WITH",
+    # COMPENSATED_BY
+    "RESCUED_BY": "COMPENSATED_BY",
+    "FUNCTIONALLY_REPLACED_BY": "COMPENSATED_BY",
+    "REDUNDANT_WITH": "COMPENSATED_BY",
+    # SUBSTRATE_OF
+    "CLEAVED_BY": "SUBSTRATE_OF",
+    "PHOSPHORYLATED_BY": "SUBSTRATE_OF",
+    "ACTED_UPON_BY": "SUBSTRATE_OF",
+    # TRANSPORTS
+    "CARRIES": "TRANSPORTS",
+    "SHUTTLES": "TRANSPORTS",
+    "TRANSLOCATES": "TRANSPORTS",
+    "MOVES": "TRANSPORTS",
+    "FERRIES": "TRANSPORTS",
+    "EXPORTS": "TRANSPORTS",
+    "IMPORTS": "TRANSPORTS",
+    # MODULATES
+    "AFFECTS": "MODULATES",
+    "ALTERS": "MODULATES",
+    "TUNES": "MODULATES",
+    "ADJUSTS": "MODULATES",
+    # LOSS_OF_FUNCTION / GAIN_OF_FUNCTION
+    "LOF": "LOSS_OF_FUNCTION",
+    "ABLATES_FUNCTION_OF": "LOSS_OF_FUNCTION",
+    "KNOCKS_OUT": "LOSS_OF_FUNCTION",
+    "INACTIVATES": "LOSS_OF_FUNCTION",
+    "GOF": "GAIN_OF_FUNCTION",
+    "CONFERS_ACTIVITY_TO": "GAIN_OF_FUNCTION",
+    "HYPERACTIVATES": "GAIN_OF_FUNCTION",
+    "CONSTITUTIVELY_ACTIVATES": "GAIN_OF_FUNCTION",
+    # SUPPORTS
+    "PROVIDES_EVIDENCE_FOR": "SUPPORTS",
+    "BACKS": "SUPPORTS",
+    "CORROBORATES": "SUPPORTS",
+    "VALIDATES": "SUPPORTS",
+    "CONFIRMS": "SUPPORTS",
+    "REINFORCES": "SUPPORTS",
+    # MENTIONS / CITES
+    "REFERENCES": "MENTIONS",
+    "DISCUSSES": "MENTIONS",
+    "DESCRIBES": "MENTIONS",
+    "NOTES": "MENTIONS",
+    "REPORTS": "MENTIONS",
+    "REFERENCES_PUBLICATION": "CITES",
+    "REFERS_TO": "CITES",
+    # INSTANCE_OF / REFINES / GENERALIZES
+    "EXAMPLE_OF": "INSTANCE_OF",
+    "IS_A": "INSTANCE_OF",
+    "TYPE_OF": "INSTANCE_OF",
+    "KIND_OF": "INSTANCE_OF",
+    "NARROWS": "REFINES",
+    "SPECIALIZES": "REFINES",
+    "BROADENS": "GENERALIZES",
+    "ABSTRACTS": "GENERALIZES",
+    # HAS_AUTHOR / HAS_KEYWORD
+    "AUTHORED_BY": "HAS_AUTHOR",
+    "WRITTEN_BY": "HAS_AUTHOR",
+    "TAGGED_WITH": "HAS_KEYWORD",
+    "INDEXED_AS": "HAS_KEYWORD",
+}
+
+
+__all__ = ["LLM_RELATION_SYNONYMS", "LLM_VALID_RELATION_TYPES"]
